ref(chunks): Extract upload logic from `upload_difs_chunked`
The `upload_difs_chunked` function contains logic specific to uploading DIFs, in addition to logic general to chunked uploads.

This change extracts the general chunked upload logic, so it can be reused for other types of chunked uploads.
