+++ src/utils/dif_upload.rs
@@ -14,8 +14,7 @@ use std::ops::Deref;
 use std::path::{Component, Path, PathBuf};
 use std::process::Command;
 use std::str;
-use std::thread;
-use std::time::{Duration, Instant};
+use std::time::Duration;
 
 use anyhow::{bail, format_err, Error, Result};
 use console::style;
@@ -33,15 +32,11 @@ use which::which;
 use zip::result::ZipError;
 use zip::{write::FileOptions, ZipArchive, ZipWriter};
 
-use crate::api::{
-    Api, AssembleDifsRequest, ChunkServerOptions, ChunkUploadCapability, ChunkedFileState,
-};
+use crate::api::{Api, ChunkServerOptions, ChunkUploadCapability};
 use crate::config::Config;
 use crate::constants::{DEFAULT_MAX_DIF_SIZE, DEFAULT_MAX_WAIT};
-use crate::utils::chunks::{
-    upload_chunks, Assemblable, BatchedSliceExt, Chunk, ChunkOptions, Chunked, ItemSize,
-    MissingObjectsInfo, ASSEMBLE_POLL_INTERVAL,
-};
+use crate::utils::chunks;
+use crate::utils::chunks::{Assemblable, BatchedSliceExt, ChunkOptions, Chunked, ItemSize};
 use crate::utils::dif::ObjectDifFeatures;
 use crate::utils::fs::{get_sha1_checksum, TempDir, TempFile};
 use crate::utils::progress::{ProgressBar, ProgressStyle};
@@ -1222,300 +1217,6 @@ fn create_il2cpp_mappings<'a>(difs: &[DifMatch<'a>]) -> Result<Vec<DifMatch<'a>>
     Ok(line_mappings)
 }
 
-/// Calls the assemble endpoint and returns the state for every `DifMatch` along
-/// with info on missing chunks.
-///
-/// The returned value contains separate vectors for incomplete DIFs and
-/// missing chunks for convenience.
-fn try_assemble<'m, T>(
-    objects: &'m [Chunked<T>],
-    options: &ChunkOptions<'_>,
-) -> Result<MissingObjectsInfo<'m, T>>
-where
-    T: AsRef<[u8]> + Assemblable,
-{
-    let api = Api::current();
-    let mut request: AssembleDifsRequest<'_> = objects.iter().collect();
-
-    if options.should_strip_debug_ids() {
-        request.strip_debug_ids();
-    }
-
-    let response =
-        api.authenticated()?
-            .assemble_difs(options.org(), options.project(), &request)?;
-
-    // We map all DIFs by their checksum, so we can access them faster when
-    // iterating through the server response below. Since the caller will invoke
-    // this function multiple times (most likely twice), this operation is
-    // performed twice with the same data. While this is redundant, it is also
-    // fast enough and keeping it here makes the `try_assemble_difs` interface
-    // nicer.
-    let difs_by_checksum = objects
-        .iter()
-        .map(|m| (m.checksum(), m))
-        .collect::<BTreeMap<_, _>>();
-
-    let mut difs = Vec::new();
-    let mut chunks = Vec::new();
-    for (checksum, ref file_response) in response {
-        let chunked_match = *difs_by_checksum
-            .get(&checksum)
-            .ok_or_else(|| format_err!("Server returned unexpected checksum"))?;
-
-        match file_response.state {
-            ChunkedFileState::Error => {
-                // One of the files could not be uploaded properly and resulted
-                // in an error. We include this file in the return value so that
-                // it shows up in the final report.
-                difs.push(chunked_match);
-            }
-            ChunkedFileState::Assembling => {
-                // This file is currently assembling. The caller will have to poll this file later
-                // until it either resolves or errors.
-                difs.push(chunked_match);
-            }
-            ChunkedFileState::NotFound => {
-                // Assembling for one of the files has not started because some
-                // (or all) of its chunks have not been found. We report its
-                // missing chunks to the caller and then continue. The caller
-                // will have to call `try_assemble_difs` again after uploading
-                // them.
-                let mut missing_chunks = chunked_match
-                    .iter_chunks()
-                    .filter(|&Chunk((c, _))| file_response.missing_chunks.contains(&c))
-                    .peekable();
-
-                // Usually every file that is NotFound should also contain a set
-                // of missing chunks. However, if we tried to upload an empty
-                // file or the server returns an invalid response, we need to
-                // make sure that this match is not included in the missing
-                // difs.
-                if missing_chunks.peek().is_some() {
-                    difs.push(chunked_match);
-                }
-
-                chunks.extend(missing_chunks);
-            }
-            _ => {
-                // This file has already finished. No action required anymore.
-            }
-        }
-    }
-
-    Ok((difs, chunks))
-}
-
-/// Concurrently uploads chunks specified in `missing_info` in batches. The
-/// batch size and number of concurrent requests is controlled by
-/// `chunk_options`.
-///
-/// This function blocks until all chunks have been uploaded.
-fn upload_missing_chunks<T>(
-    missing_info: &MissingObjectsInfo<'_, T>,
-    chunk_options: &ChunkServerOptions,
-) -> Result<()> {
-    let (objects, chunks) = missing_info;
-
-    // Chunks might be empty if errors occurred in a previous upload. We do
-    // not need to render a progress bar or perform an upload in this case.
-    if chunks.is_empty() {
-        return Ok(());
-    }
-
-    let progress_style = ProgressStyle::default_bar().template(&format!(
-        "{} Uploading {} missing debug information file{}...\
-         \n{{wide_bar}}  {{bytes}}/{{total_bytes}} ({{eta}})",
-        style(">").dim(),
-        style(objects.len().to_string()).yellow(),
-        if objects.len() == 1 { "" } else { "s" }
-    ));
-
-    upload_chunks(chunks, chunk_options, progress_style)?;
-
-    println!(
-        "{} Uploaded {} missing debug information {}",
-        style(">").dim(),
-        style(objects.len().to_string()).yellow(),
-        match objects.len() {
-            1 => "file",
-            _ => "files",
-        }
-    );
-
-    Ok(())
-}
-
-/// Renders the given detail string to the command line. If the `detail` is
-/// either missing or empty, the optional fallback will be used.
-fn render_detail(detail: &Option<String>, fallback: Option<&str>) {
-    let mut string = match *detail {
-        Some(ref string) => string.as_str(),
-        None => "",
-    };
-
-    if string.is_empty() {
-        if let Some(fallback) = fallback {
-            string = fallback;
-        }
-    }
-
-    for line in string.lines() {
-        if !line.is_empty() {
-            println!("        {}", style(line).dim());
-        }
-    }
-}
-
-/// Polls the assemble endpoint until all DIFs have either completed or errored. Returns a list of
-/// `DebugInfoFile`s that have been created successfully and also prints a summary to the user.
-///
-/// This function assumes that all chunks have been uploaded successfully. If there are still
-/// missing chunks in the assemble response, this likely indicates a bug in the server.
-fn poll_assemble<T>(
-    chunked_objects: &[&Chunked<T>],
-    options: &ChunkOptions<'_>,
-) -> Result<(Vec<DebugInfoFile>, bool)>
-where
-    T: Display + Assemblable,
-{
-    let progress_style = ProgressStyle::default_bar().template(
-        "{prefix:.dim} Processing files...\
-         \n{wide_bar}  {pos}/{len}",
-    );
-
-    let api = Api::current();
-    let pb = ProgressBar::new(chunked_objects.len());
-    pb.set_style(progress_style);
-    pb.set_prefix(">");
-
-    let assemble_start = Instant::now();
-
-    let mut request: AssembleDifsRequest<'_> = chunked_objects.iter().copied().collect();
-    if options.should_strip_debug_ids() {
-        request.strip_debug_ids();
-    }
-
-    let response = loop {
-        let response =
-            api.authenticated()?
-                .assemble_difs(options.org(), options.project(), &request)?;
-
-        let chunks_missing = response
-            .values()
-            .any(|r| r.state == ChunkedFileState::NotFound);
-
-        if chunks_missing {
-            return Err(format_err!(
-                "Some uploaded files are now missing on the server. Please retry by running \
-        `sentry-cli upload-dif` again. If this problem persists, please report a bug.",
-            ));
-        }
-
-        // Poll until there is a response, unless the user has specified to skip polling. In
-        // that case, we return the potentially partial response from the server. This might
-        // still contain a cached error.
-        if !options.should_wait() {
-            break response;
-        }
-
-        if assemble_start.elapsed() > options.max_wait() {
-            break response;
-        }
-
-        let pending = response
-            .iter()
-            .filter(|&(_, r)| r.state.is_pending())
-            .count();
-
-        pb.set_position((chunked_objects.len() - pending) as u64);
-
-        if pending == 0 {
-            break response;
-        }
-
-        thread::sleep(ASSEMBLE_POLL_INTERVAL);
-    };
-
-    pb.finish_and_clear();
-    if response.values().any(|r| r.state.is_pending()) {
-        println!("{} File upload complete:\n", style(">").dim());
-    } else {
-        println!("{} File processing complete:\n", style(">").dim());
-    }
-
-    let (errors, mut successes): (Vec<_>, _) = response
-        .into_iter()
-        .partition(|(_, r)| r.state.is_err() || options.should_wait() && r.state.is_pending());
-
-    // Print a summary of all successes first, so that errors show up at the
-    // bottom for the user
-    successes.sort_by_key(|(_, success)| {
-        success
-            .dif
-            .as_ref()
-            .map(|x| x.object_name.as_str())
-            .unwrap_or("")
-            .to_owned()
-    });
-
-    let objects_by_checksum: BTreeMap<_, _> =
-        chunked_objects.iter().map(|m| (m.checksum(), m)).collect();
-
-    for &(checksum, ref success) in &successes {
-        // Silently skip all OK entries without a "dif" record since the server
-        // will always return one.
-        if let Some(ref dif) = success.dif {
-            // Files that have completed processing will contain a `dif` record
-            // returned by the server. Use this to show detailed information.
-            println!(
-                "  {:>7} {} ({}; {}{})",
-                style("OK").green(),
-                style(&dif.id()).dim(),
-                dif.object_name,
-                dif.cpu_name,
-                dif.data.kind.map(|c| format!(" {c:#}")).unwrap_or_default()
-            );
-
-            render_detail(&success.detail, None);
-        } else if let Some(object) = objects_by_checksum.get(&checksum) {
-            // If we skip waiting for the server to finish processing, there
-            // are pending entries. We only expect results that have been
-            // uploaded in the first place, so we can skip everything else.
-            println!("  {:>8} {}", style("UPLOADED").yellow(), object);
-        }
-        // All other entries will be in the `errors` list.
-    }
-
-    // Print a summary of all errors at the bottom.
-    let mut errored = vec![];
-    for (checksum, error) in errors {
-        let object = objects_by_checksum
-            .get(&checksum)
-            .ok_or_else(|| format_err!("Server returned unexpected checksum"))?;
-        errored.push((object, error));
-    }
-    errored.sort_by_key(|x| x.0.name());
-
-    let has_errors = !errored.is_empty();
-    for (object, error) in errored {
-        let fallback = match error.state {
-            ChunkedFileState::Assembling => Some("The file is still processing and not ready yet"),
-            ChunkedFileState::NotFound => Some("The file could not be saved"),
-            _ => Some("An unknown error occurred"),
-        };
-
-        println!("  {:>7} {}", style("ERROR").red(), object.name());
-        render_detail(&error.detail, fallback);
-    }
-
-    // Return only successful uploads
-    Ok((
-        successes.into_iter().filter_map(|(_, r)| r.dif).collect(),
-        has_errors,
-    ))
-}
-
 /// Uploads debug info files using the chunk-upload endpoint.
 fn upload_difs_chunked(
     options: DifUpload,
@@ -1549,23 +1250,7 @@ fn upload_difs_chunked(
     })?;
 
     let options = options.into_chunk_options(chunk_options);
-
-    // Upload missing chunks to the server and remember incomplete difs
-    let missing_info = try_assemble(&chunked, &options)?;
-    upload_missing_chunks(&missing_info, options.server_options())?;
-
-    // Only if DIFs were missing, poll until assembling is complete
-    let (missing_difs, _) = missing_info;
-    if !missing_difs.is_empty() {
-        poll_assemble(&missing_difs, &options)
-    } else {
-        println!(
-            "{} Nothing to upload, all files are on the server",
-            style(">").dim()
-        );
-
-        Ok((Default::default(), false))
-    }
+    chunks::upload_chunked_objects(&chunked, options)
 }
 
 /// Returns debug files missing on the server.
